+++ toolsets/oracle-operator/types/flags.go
@@ -2,7 +2,6 @@ package types
 
 // default Flags
 const (
-	FlagLogLevel      = "log_level"
-	FlagQueryEndpoint = "query_endpoint"
-	FlagQueryMethod   = "query_method"
+	FlagLogLevel   = "log_level"
+	FlagConfigFile = "config"
 )