+++ x/gov/simulation/operations.go
@@ -341,7 +341,11 @@ func randomProposalID(r *rand.Rand, k keeper.Keeper,
 	switch {
 	case proposalID > initialProposalID:
 		// select a random ID between [initialProposalID, proposalID]
-		proposalID = uint64(simulation.RandIntBetween(r, int(initialProposalID), int(proposalID)))
+		diff := int64(proposalID - initialProposalID)
+		if diff < 0 {
+			diff *= -1
+		}
+		proposalID = uint64(r.Int63n(diff)) + initialProposalID
 
 	default:
 		// This is called on the first call to this funcion