+++ x/oracle/abci.go
@@ -1,6 +1,8 @@
 package oracle
 
 import (
+	"strconv"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/certikfoundation/shentu/x/oracle/internal/keeper"
@@ -22,20 +24,27 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) {
 			continue
 		}
 
+		if err := k.DistributeBounty(ctx, task); err != nil {
+			// TODO
+			continue
+		}
+
 		ctx.EventManager().EmitEvent(
 			sdk.NewEvent(
 				"aggregate_task",
 				sdk.NewAttribute("contract", task.Contract),
 				sdk.NewAttribute("function", task.Function),
-				sdk.NewAttribute("result", task.Result.String()),
+				sdk.NewAttribute("begin_block_height", strconv.FormatInt(task.BeginBlock, 10)),
+				sdk.NewAttribute("bounty", task.Bounty.String()),
+				sdk.NewAttribute("description", task.Description),
 				sdk.NewAttribute("expiration", task.Expiration.String()),
+				sdk.NewAttribute("creator", task.Creator.String()),
+				sdk.NewAttribute("responses", task.Responses.String()),
+				sdk.NewAttribute("result", task.Result.String()),
+				sdk.NewAttribute("end_block_height", strconv.FormatInt(task.ClosingBlock, 10)),
+				sdk.NewAttribute("status", task.Status.String()),
 			),
 		)
-
-		if err := k.DistributeBounty(ctx, task); err != nil {
-			// TODO
-			continue
-		}
 	}
 	k.DeleteClosingTaskIDs(ctx, ctx.BlockHeight())
 }